requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
Werkzeug==3.0.1
email-validator==2.2.0
openpyxl==3.1.2
gunicorn==21.2.0
waitress==3.0.0
requests==2.31.0
deep-translator==1.11.4

